06-lab-complete/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
redis==5.0.1
pydantic==2.9.0
pydantic-settings==2.5.2
requests==2.31.0
psutil==6.0.0
PyJWT==2.9.0
